+++ client/src/input/Input.js
@@ -132,6 +132,7 @@ export default Garnish.Base.extend({
         fieldLayoutId: blockType.getFieldLayoutId(),
         name: blockType.getName(),
         handle: blockType.getHandle(),
+        description: blockType.getDescription(),
         maxBlocks: blockType.getMaxBlocks(),
         maxSiblingBlocks: blockType.getMaxSiblingBlocks(),
         maxChildBlocks: blockType.getMaxChildBlocks(),