+++ src/services/BlockTypes.php
@@ -201,6 +201,7 @@ class BlockTypes extends Component
         $record->fieldLayoutId = $blockType->fieldLayoutId;
         $record->name = $blockType->name;
         $record->handle = $blockType->handle;
+        $record->description = $blockType->description;
         $record->sortOrder = $blockType->sortOrder;
         $record->maxBlocks = $blockType->maxBlocks;
         $record->maxSiblingBlocks = $blockType->maxSiblingBlocks;
@@ -415,6 +416,7 @@ class BlockTypes extends Component
             $record->groupId = $groupId;
             $record->name = $data['name'];
             $record->handle = $data['handle'];
+            $record->description = $data['description'];
             $record->sortOrder = $data['sortOrder'];
             $record->maxBlocks = $data['maxBlocks'];
             $record->maxSiblingBlocks = $data['maxSiblingBlocks'] ?? 0;
@@ -430,6 +432,7 @@ class BlockTypes extends Component
             $blockType->groupId = $groupId;
             $blockType->name = $data['name'];
             $blockType->handle = $data['handle'];
+            $blockType->description = $data['description'];
             $blockType->sortOrder = $data['sortOrder'];
             $blockType->maxBlocks = $data['maxBlocks'];
             $blockType->maxSiblingBlocks = $data['maxSiblingBlocks'] ?? 0;
@@ -636,6 +639,7 @@ class BlockTypes extends Component
                 'groupId',
                 'name',
                 'handle',
+                'description',
                 'maxBlocks',
                 'maxSiblingBlocks',
                 'maxChildBlocks',